A couple small tweeks.
Shuffles the embedding for each prompt in the batch individually.  The embedding does remain consistent within each prompt, in case there is a prompt like: "a painting of * smiling beside *"
`shuffle_off` no longer performs indexing when `num_vectors` is the same as the max number of vectors.  Probably doesn't matter, but it likely gets rid of some needless indirection.